boss rize bug

+++ Game/Scripts/Utils/GlobalsConfigs.cs
@@ -81,8 +81,8 @@ namespace Game.Scripts.Utils
         public int SpecialWaveEnemyCount { get; set; }
         public float TimeBetweenWaves { get; set; }
         public int BasicWavesBeforeSpecial { get; set; }
-        public int TimeUntilFirstWave { get; set; }
-        public int TimeUntilGameEnds { get; set; }
+        public float TimeUntilFirstWave { get; set; }
+        public float TimeUntilGameEnds { get; set; }
     }
 
     [Serializable]
